requirements.txt
aniso8601
beautifulsoup4
certifi
cffi
charset-normalizer
click
cryptography
dataclasses
Deprecated
dnspython
ecdsa
email-validator
Flask
Flask-Cors
Flask-JWT-Extended
flask-mongoengine
Flask-RESTful
Flask-WTF
gunicorn
idna
importlib-metadata
itsdangerous
jdatetime
Jinja2
MarkupSafe
mongoengine
Pillow
pyasn1
pycparser
PyGithub
PyJWT
pymongo
PyNaCl
python-jose
python-opengraph-jaywink
pytz
requests
rsa
six
soupsieve
typing-extensions
urllib3
uWSGI
Werkzeug
wrapt
WTForms
zipp
